+++ src/generators/java/mekanism/generators/common/content/turbine/SynchronizedTurbineData.java
@@ -55,11 +55,11 @@ public class SynchronizedTurbineData extends SynchronizedData<SynchronizedTurbin
 
     public Coord4D complex;
 
-    public int lastSteamInput;
-    public int newSteamInput;
+    public long lastSteamInput;
+    public long newSteamInput;
 
     public int clientDispersers;
-    public int clientFlow;
+    public long clientFlow;
     public float clientRotation;
 
     public SynchronizedTurbineData(TileEntityTurbineCasing tile) {
