+++ src/main/java/mekanism/client/jei/chemical/InfusionStackRenderer.java
@@ -10,7 +10,7 @@ public class InfusionStackRenderer extends ChemicalStackRenderer<InfuseType, Inf
         super(FluidAttributes.BUCKET_VOLUME, TooltipMode.ITEM_LIST, TEX_WIDTH, TEX_HEIGHT, null);
     }
 
-    public InfusionStackRenderer(int capacityMb, int width, int height) {
+    public InfusionStackRenderer(long capacityMb, int width, int height) {
         super(capacityMb, TooltipMode.SHOW_AMOUNT_NO_UNITS, width, height, null);
     }
 }
